+++ src/components/DataTable/index.js
@@ -2,19 +2,20 @@ import React, { useState, useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link } from 'react-router-dom';
 import { saveAs } from 'file-saver';
-import { format } from 'date-fns';
 
 import { 
 	Card, Row, Col,
 	CardFooter, CardHeader, CardBody,
-	Button, Input, InputGroup,
-	InputGroupAddon, InputGroupText,
-	Dropdown, DropdownToggle,
+	Button, Dropdown, DropdownToggle,
 	DropdownMenu, DropdownItem, Container
 } from 'reactstrap';
 
 import Table from './Table';
 import Pagination from '../Pagination';
+import LimitDropdown from './LimitDropdown';
+import { DownloadButton, CreateButton } from './Buttons';
+import Search from './Search';
+import Sort from './Sort';
 // import CustomDropdownButton from './CustomDropdownButton'; DON'T REMOVE YET. IT MAY BE USEFUL ON REFACTORING DATATABLE
 
 import { CellContentLoader } from '../ContentLoader';
@@ -34,8 +35,7 @@ export function DataTable ({
 	contentLoader = true
 	}) {
 	const [filterValue, setFilterValue] = useState('');
-	const [isSortOpen, setSortDropdown] = useState(false);
-	const [isLimitOpen, setLimitDropdown] = useState(false);
+
 	const timeoutRef = useRef(null);
 	
 	const { t } = useTranslation();
@@ -51,22 +51,6 @@ export function DataTable ({
 		if (onSearch) onSearch(filterValue);
 		}, 500);
 	}, [filterValue]);
-
-	const downloadHandler = () => {
-		const list = onDownload();
-		let csv = headers.map(header => header.name).join(',') + "\n" + 
-			list.map(item => headers.map(header => {
-				if (header.customComponent) {
-					if (header.customComponent.onDownload)
-						return header.customComponent.onDownload(item, header).replace(',', ';');
-					return '-';
-				}
-				return JSON.stringify(item[header.key])?.replace(',', ';');
-			}).join(',')).join('\n');
-		let blob = new Blob([csv], {type: "text/csv;charset=utf-8"});
-		let name = title?.text || "";
-		saveAs(blob, `${name.replace(' ', '_')}_${format(new Date, 'dd-MM-yyyy')}.csv`);
-	}
 
 	return (
 		<Row>
@@ -98,78 +82,45 @@ export function DataTable ({
 						}
 						{buttonWithAuthz && <ButtonWithAuthz {...buttonWithAuthz} className="float-right ml-3 data-table-button-with-authz"/>}
 						{createButton &&
-							<div className="float-right ml-3 data-table-create-button">
-								<Link to={{ pathname: createButton.pathname }}>
-									<Button tag="span" size="sm">
-										{createButton.icon && 
-											<span className="pr-1">
-												{typeof createButton.icon === 'string' ? <i className={createButton.icon}></i> : createButton.icon}
-											</span>
-										}
-										{createButton.text}
-									</Button>
-								</Link>
-							</div>
+							<CreateButton 
+								createButton={createButton}
+							/>
 						}
 						{onDownload &&
 							<div className="float-right ml-3 data-table-download-button">
-								<Button tag="span" size="sm" onClick={downloadHandler} >
-									<i className="cil-arrow-bottom"></i>
-									Download
-								</Button>
+								<DownloadButton
+									onDownload={onDownload}
+									headers={headers}
+									title={title}
+								/>
 							</div>
 						}
 						{sort && 
 							<div className="float-right ml-3 data-table-sort">
-								<Dropdown
-									isOpen={isSortOpen}
-									toggle={() => setSortDropdown(prev => !prev)}
-								>
-									<DropdownToggle size="sm" caret>
-										{sort.icon && <i className={`${sort.icon} mr-1`}></i>}
-										{sort.title}
-									</DropdownToggle>
-									<DropdownMenu>
-									{
-										sort.items.map((item, idx) => (
-											<DropdownItem onClick={() => sort.onClick(item.value)} key={idx}>
-												{item.icon && <i className={`${item.icon} mr-1`}></i>}
-												{item.name}
-											</DropdownItem>
-										))
-									}
-									</DropdownMenu>
-								</Dropdown>
+								<Sort 
+									sort={sort} 
+								/>
 							</div>
 						}
 						{search && 
 							<div className="float-right ml-3 data-table-search">
-								<InputGroup>
-								{search.icon && 
-									<InputGroupAddon addonType="prepend">
-									<InputGroupText><i className={search.icon}></i></InputGroupText>
-									</InputGroupAddon>
-								}
-								<Input
-									value={filterValue}
-									onChange={e => setFilterValue(e.target.value)}
-									placeholder={search.placeholder}
-									type="text"
-									bsSize="sm"
+								<Search 
+									search={search}
+									filterValue={filterValue}
+									setFilterValue={setFilterValue}
 								/>
-								</InputGroup>
 							</div>
 						}
 					</CardHeader>
 
 					<CardBody className="data-table-card-body">
 						{customCardBodyComponent}
-						<Table
+						{!isLoading && <Table
 							data={data.length > limit ? data.slice(0, limit) : data}
 							headers={headers}
 							rowStyle={customRowStyle}
 							rowClassName={customRowClassName}
-						/>
+						/>}
 
 						{isLoading && contentLoader && <CellContentLoader cols={headers.length} rows={limit ?? 5} /> }
 
@@ -207,20 +158,13 @@ export function DataTable ({
 
 						{setLimit &&
 							<Col>
-								<Dropdown
-									isOpen={isLimitOpen}
-									toggle={() => setLimitDropdown(prev => !prev)}
-									className="float-right"
-								>
-									<DropdownToggle caret size="sm">
-										{t(translationRoute ? `${translationRoute}|Limit` : "Limit")}: {limit}
-									</DropdownToggle>
-									<DropdownMenu>
-									{
-										limitValues.map((value, idx) => <DropdownItem onClick={() => { setPage(1); setLimit(value); }} key={idx}>{value}</DropdownItem>)
-									}
-									</DropdownMenu>
-								</Dropdown>
+								<LimitDropdown 
+									translationRoute={translationRoute}
+									limit={limit}
+									setLimit={setLimit}
+									setPage={setPage}
+									limitValues={limitValues}
+								/>
 							</Col>
 						}
 						</Row>
